Removing and merging single method files
