+++ athena-gcs/src/main/java/com/amazonaws/athena/connectors/gcs/GcsFieldResolver.java
@@ -21,7 +21,10 @@ package com.amazonaws.athena.connectors.gcs;
 
 import com.amazonaws.athena.connector.lambda.data.FieldResolver;
 import org.apache.arrow.vector.types.Types;
+import org.apache.arrow.vector.types.pojo.ArrowType;
 import org.apache.arrow.vector.types.pojo.Field;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.util.Map;
 
@@ -33,6 +36,7 @@ import java.util.Map;
 public class GcsFieldResolver
         implements FieldResolver
 {
+    private static final Logger LOGGER = LoggerFactory.getLogger(GcsFieldResolver.class);
     protected static final FieldResolver DEFAULT_FIELD_RESOLVER = new GcsFieldResolver();
 
     private GcsFieldResolver() {}
@@ -43,12 +47,68 @@ public class GcsFieldResolver
     {
         Types.MinorType minorType = Types.getMinorTypeForArrowType(field.getType());
         if (minorType == Types.MinorType.LIST) {
-            return TypeUtils.coerce(field, ((Map<String, Object>) value).get(field.getName()));
+            return coerce(field, ((Map<String, Object>) value).get(field.getName()));
         }
         else if (value instanceof Map) {
             Object rawVal = ((Map<String, Object>) value).get(field.getName());
-            return TypeUtils.coerce(field, rawVal);
+            return coerce(field, rawVal);
         }
         throw new RuntimeException("Expected LIST or Document type but found " + minorType);
+    }
+
+    /**
+     * Allows for coercing types in the event that schema has evolved or there were other data issues.
+     *
+     * @param field The field that we are coercing the value into.
+     * @param origVal The value to coerce
+     * @return The coerced value.
+     * coercions in the future as a way of dealing with schema evolution.
+     */
+    public static Object coerce(Field field, Object origVal)
+    {
+        if (origVal == null) {
+            return origVal;
+        }
+        LOGGER.info("TypeUtils.coerce(Field, Object) :: FQCN is {}", origVal.getClass().getName());
+        ArrowType arrowType = field.getType();
+        Types.MinorType minorType = Types.getMinorTypeForArrowType(arrowType);
+
+        switch (minorType) {
+            case VARCHAR:
+                if (origVal instanceof String) {
+                    return origVal;
+                }
+                else {
+                    return String.valueOf(origVal);
+                }
+            case FLOAT8:
+                if (origVal instanceof Integer) {
+                    return (double) (int) origVal;
+                }
+                else if (origVal instanceof Float) {
+                    return (double) (float) origVal;
+                }
+                return origVal;
+            case FLOAT4:
+                if (origVal instanceof Integer) {
+                    return (float) (int) origVal;
+                }
+                else if (origVal instanceof Double) {
+                    return ((Double) origVal).floatValue();
+                }
+                return origVal;
+            case INT:
+                if (origVal instanceof Float) {
+                    return ((Float) origVal).intValue();
+                }
+                else if (origVal instanceof Double) {
+                    return ((Double) origVal).intValue();
+                }
+                return origVal;
+            case DATEMILLI:
+                return origVal.toString();
+            default:
+                return origVal;
+        }
     }
 }
